requirements.txt
apache-airflow
boto3
botocore
str2bool
pandas
lightgbm
xgboost
scikit-learn
joblib
